requirements.txt
appnope==0.1.0
backcall==0.1.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.3
idna==2.8
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
oauthlib==3.0.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
Pygments==2.4.0
python-dateutil==2.8.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.2
wcwidth==0.1.7
